Extract sortedKeys helper to shared helpers.go
Move the sortedKeys utility from cancel_order.go to helpers.go since
it's used by both cancel_order.go and create_collection.go. Prevents
confusion about cross-file dependencies.
