+++ connectors/shopify/cancel_order.go
@@ -5,8 +5,6 @@ import (
 	"encoding/json"
 	"fmt"
 	"net/http"
-	"sort"
-	"strings"
 
 	"github.com/supersuit-tech/permission-slip-web/connectors"
 )
@@ -46,18 +44,6 @@ func (p *cancelOrderParams) validate() error {
 		}
 	}
 	return nil
-}
-
-// sortedKeys returns the keys of a map[string]bool as a sorted,
-// comma-separated string. Used in validation error messages so they
-// stay in sync with the allowed-values maps automatically.
-func sortedKeys(m map[string]bool) string {
-	keys := make([]string, 0, len(m))
-	for k := range m {
-		keys = append(keys, k)
-	}
-	sort.Strings(keys)
-	return strings.Join(keys, ", ")
 }
 
 // Execute cancels a Shopify order. This action is irreversible.
